+++ py/desispec/preproc.py
@@ -9,6 +9,7 @@ import scipy.interpolate
 from pkg_resources import resource_exists, resource_filename
 
 from scipy import signal
+from scipy.ndimage.filters import median_filter
 
 from desispec.image import Image
 from desispec import cosmics
@@ -25,23 +26,16 @@ from desispec.maskedmedian import masked_median
 from desispec.image_model import compute_image_model
 from desispec.util import header2night
 
-# log = get_logger()
-
 def get_amp_ids(header):
     '''
-    Return list of amp names ['A','B','C','D'] or ['1','2','3','4']
-    based upon header keywords
+    Return list of amp names based upon header keywords
     '''
-    if 'CCDSECA' in header:
-        amp_ids = ['A', 'B', 'C', 'D']
-    elif 'CCDSEC1' in header:
-        amp_ids = ['1', '2', '3', '4']
-    else:
-        log = get_logger()
-        message = "No CCDSECA or CCDSEC1; Can't derive amp names from header"
-        log.fatal(message)
-        raise KeyError(message)
-
+    amp_ids = []
+    for a in ['A', 'B', 'C', 'D', '1', '2', '3', '4'] :
+        if 'BIASSEC'+a in header :
+            amp_ids.append(a)
+    if len(amp_ids)==0 :
+        raise KeyError("No keyword BIASSECX with X in A,B,C,D,1,2,3,4 in header")
     return amp_ids
 
 def _parse_sec_keyword(value):
@@ -146,21 +140,18 @@ def subtract_peramp_overscan(image, hdr):
     has more complex support for row-by-row, col-overscan, etc.)
     """
     amp_ids = get_amp_ids(hdr)
-    n0=image.shape[0]//2
-    n1=image.shape[1]//2
-
     for a,amp in enumerate(amp_ids) :
-        ii = parse_sec_keyword(hdr['BIASSEC'+amp])
+        ii=parse_sec_keyword(hdr['BIASSEC'+amp])
+        s0,s1=ii[0],ii[1]
+        for k in ["DATASEC","PRESEC","ORSEC","PRRSEC"] :
+            if k+amp in hdr :
+                t0,t1=parse_sec_keyword(hdr[k+amp])
+                s0 = slice(min(s0.start,t0.start),max(s0.stop,t0.stop))
+                s1 = slice(min(s1.start,t1.start),max(s1.stop,t1.stop))
         overscan_image = image[ii].copy()
         overscan,rdnoise = calc_overscan(overscan_image)
-        if ii[0].start < n0 and ii[1].start < n1 :
-            image[:n0,:n1] -= overscan
-        elif ii[0].start < n0 and ii[1].start >= n1 :
-            image[:n0,n1:] -= overscan
-        elif ii[0].start >= n0 and ii[1].start < n1 :
-            image[n0:,:n1] -= overscan
-        elif ii[0].start >= n0 and ii[1].start >= n1 :
-            image[n0:,n1:] -= overscan
+        image[s0,s1] -= overscan
+
 
 def _savgol_clipped(data, window=15, polyorder=5, niter=0, threshold=3.):
     """
@@ -654,44 +645,63 @@ def preproc(rawimage, header, primary_header, bias=True, dark=True, pixflat=True
                 o,r = calc_overscan(raw_overscan_squared[row])
                 overscan_row[row] = raw_overscan_row[row] - o
 
+        kk = parse_sec_keyword(header['CCDSEC'+amp])
+
         # Now remove the overscan_col
         nrows=raw_overscan_col.shape[0]
         log.info("nrows in overscan=%d"%nrows)
         overscan_col = np.zeros(nrows)
         rdnoise  = np.zeros(nrows)
-        if (cfinder and cfinder.haskey('OVERSCAN'+amp) and cfinder.value("OVERSCAN"+amp).upper()=="PER_ROW") or overscan_per_row:
-            log.info("Subtracting overscan per row for amplifier %s of camera %s"%(amp,camera))
-            for j in range(nrows) :
-                if np.isnan(np.sum(overscan_col[j])) :
-                    log.warning("NaN values in row %d of overscan of amplifier %s of camera %s"%(j,amp,camera))
-                    continue
-                o,r =  calc_overscan(raw_overscan_col[j])
-                #log.info("%d %f %f"%(j,o,r))
-                overscan_col[j]=o
-                rdnoise[j]=r
-        else :
+        for j in range(nrows) :
+            if np.isnan(np.sum(overscan_col[j])) :
+                log.warning("NaN values in row %d of overscan of amplifier %s of camera %s"%(j,amp,camera))
+                continue
+            o,r =  calc_overscan(raw_overscan_col[j])
+            overscan_col[j]=o
+            rdnoise[j]=r
+        # median filter to futher reduce the noise
+        med_overscan_col = median_filter(overscan_col,7)
+        # range of variation of overscan
+        margin=50
+        overscan_step = np.max(med_overscan_col[margin:-margin])-np.min(med_overscan_col[margin:-margin])
+        header['OSTEP'+amp] = (overscan_step,'ADUs (max-min of median overscan per row)')
+        log.info("Overscan max-min per row (OSTEP) for amplifier %s of camera %s = %.2f ADU"%(amp,camera,overscan_step))
+        if overscan_step <  2 : # tuned to trig on the worst few
             log.info("Subtracting average overscan for amplifier %s of camera %s"%(amp,camera))
             o,r =  calc_overscan(raw_overscan_col)
-            overscan_col += o
-            rdnoise  += r
-            if bias is not False :
-                # the master bias noise is already in the raw data
-                # (because we already subtracted the bias)
-                # so we only need to add the quadratic difference of the master bias read noise
-                # between the active region and the overscan columns
-                jj = parse_sec_keyword(header['DATASEC'+amp])
-                o,biasnoise_datasec = calc_overscan(bias[jj])
-                o,biasnoise_ovcol   = calc_overscan(bias[ov_col])
-                new_rdnoise         = np.sqrt(rdnoise**2+biasnoise_datasec**2-biasnoise_ovcol**2)
-                log.info("Master bias noise for AMP %s = %4.3f ADU, rdnoise %4.3f -> %4.3f ADU"%(amp,biasnoise_datasec,np.mean(rdnoise),np.mean(new_rdnoise)))
-                rdnoise = new_rdnoise
+            # replace by single value
+            overscan_col = np.repeat(o,nrows)
+            rdnoise  = np.repeat(r,nrows)
+            header['OMETH'+amp]=("AVERAGE","use average overscan")
+        else :
+            header['OMETH'+amp]=("PER_ROW","use average overscan per row")
+            log.info("Subtracting overscan per row for amplifier %s of camera %s"%(amp,camera))
+
+            # The threshold of 5 ADUs discards 20% of the r8-A data
+            # from Oct 2021. But it is necessary to discard some
+            # exposures with OSTEPA>=8 ADU where bias variation
+            # residuals were the cause of some bad redshifts.
+            if overscan_step > 5. :
+                mask[kk] |= ccdmask.BADREADNOISE
+                log.warning("OSTEP={} is large for amplifier {}, set ccdmask.BADREADNOISE bit mask".format(overscan_step,amp))
+
+        if bias is not False :
+            # the master bias noise is already in the raw data
+            # (because we already subtracted the bias)
+            # so we only need to add the quadratic difference of the master bias read noise
+            # between the active region and the overscan columns
+            jj = parse_sec_keyword(header['DATASEC'+amp])
+            o,biasnoise_datasec = calc_overscan(bias[jj])
+            o,biasnoise_ovcol   = calc_overscan(bias[ov_col])
+            new_rdnoise         = np.sqrt(rdnoise**2+biasnoise_datasec**2-biasnoise_ovcol**2)
+            log.info("Master bias noise for AMP %s = %4.3f ADU, rdnoise %4.3f -> %4.3f ADU"%(amp,biasnoise_datasec,np.mean(rdnoise),np.mean(new_rdnoise)))
+            rdnoise = new_rdnoise
 
         rdnoise *= gain
         median_rdnoise  = np.median(rdnoise)
         median_overscan = np.median(overscan_col)
         log.info("Median rdnoise and overscan= %f %f"%(median_rdnoise,median_overscan))
 
-        kk = parse_sec_keyword(header['CCDSEC'+amp])
         for j in range(nrows) :
             readnoise[kk][j] = rdnoise[j]
 
@@ -879,8 +889,8 @@ def preproc(rawimage, header, primary_header, bias=True, dark=True, pixflat=True
     ivar = np.zeros(var.shape)
     ivar[var>0] = 1.0 / var[var>0]
 
-    #- Ridiculously high readnoise is bad
-    mask[readnoise>100] |= ccdmask.BADREADNOISE
+    #- High readnoise is bad
+    mask[readnoise>10] |= ccdmask.BADREADNOISE
 
     if bkgsub :
         bkg = _background(image,header)